Refactoring eww. Sorry in advance
