+++ app/src/main/AndroidManifest.xml
@@ -21,19 +21,16 @@
         android:theme="@style/Theme.QArmy"
         tools:targetApi="31">
         <activity
-            android:name=".UserProfileActivity"
+            android:name=".UI.profile.UserProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".RegistrationActivity"
+            android:name=".UI.profile.RegistrationActivity"
             android:exported="true" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".UI.profile.EditProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".EditProfileActivity"
-            android:exported="false" />
-        <activity
-            android:name=".MainActivity"
+            android:name=".UI.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -42,7 +39,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".CaptureAct"
+            android:name=".UI.CaptureAct"
             android:configChanges="orientation|keyboardHidden"
             android:exported="true"
             android:screenOrientation="portrait"
