--- app/src/main/java/com/example/QArmy/QrVisual.java
@@ -1,4 +1,0 @@
-package com.example.QArmy;
-
-public class QrVisual {
-}
