+++ app/src/main/java/com/example/QArmy/UI/MapFragment.java
@@ -1,4 +1,4 @@
-package com.example.QArmy;
+package com.example.QArmy.UI;
 
 import android.content.Context;
 import android.content.pm.PackageManager;
