+++ app/src/main/java/com/example/QArmy/UI/profile/EditProfileActivity.java
@@ -1,4 +1,4 @@
-package com.example.QArmy;
+package com.example.QArmy.UI.profile;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -11,6 +11,7 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.example.QArmy.R;
 import com.example.QArmy.model.User;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
