+++ app/src/main/java/com/example/QArmy/UI/profile/UserProfileActivity.java
@@ -1,17 +1,14 @@
-package com.example.QArmy;
+package com.example.QArmy.UI.profile;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.Handler;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.QArmy.R;
 import com.example.QArmy.model.User;
