+++ app/src/main/java/com/example/QArmy/model/Image.java
@@ -1,0 +1,4 @@
+package com.example.QArmy.model;
+
+public class Image {
+}
