+++ app/src/main/java/com/example/QArmy/model/PlayerProfile.java
@@ -1,0 +1,4 @@
+package com.example.QArmy.model;
+
+public class PlayerProfile {
+}
