+++ app/src/main/java/com/example/QArmy/model/QRList.java
@@ -10,8 +10,10 @@
  * Sources:
  */
 
-package com.example.QArmy;
+package com.example.QArmy.model;
 
+import com.example.QArmy.TModel;
+import com.example.QArmy.TView;
 import com.example.QArmy.model.QRCode;
 
 import java.util.ArrayList;
