+++ app/src/main/res/layout/fragment_home.xml
@@ -18,7 +18,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
-        android:name="com.example.QArmy.SummaryFragment"
+        android:name="com.example.QArmy.UI.qrcodes.SummaryFragment"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         tools:layout="@layout/fragment_summary" />
