+++ app/src/main/res/layout/fragment_map.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".UI.MainActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
