+++ app/src/main/res/layout/fragment_summary.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:background="@color/black"
     android:orientation="horizontal"
-    tools:context=".SummaryFragment">
+    tools:context=".UI.qrcodes.SummaryFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="wrap_content"
